Fixed all findings specified in new outbound single branch
Made the large functions into classes, moved the event listeners for the
bridge, fixed naming conventions, etc.
