+++ app.js
@@ -17,14 +17,6 @@ if (confFile = process.argv[2]) {
 }
 
 /**
- * Checks if the argument tied to the StasisStart event is dialed (not inbound)
- * @param {String} argument - The argument (either an extension # or dialed)
- */
-function isDialed(argument) {
-  return argument === 'dialed';
-}
-
-/**
  * Waits for a StasisStart event before going into the main SLA module
  * @param {Object} client - Object that contains information from the ARI
  *   connection.
@@ -32,10 +24,9 @@ function isDialed(argument) {
 function clientLoaded (client) {
   client.start('sla');
   client.on('StasisStart', function(event, channel) {
-    if (!isDialed(event.args[0])) {
+    if (event.args[0] !== 'dialed') {
       var extension = event.args[0];
       sla(client, confFile, channel, extension)
-        .then(console.log)
         .catch(errHandler)
         .done();
     }
