+++ lib/states.js
@@ -1,0 +1,8 @@
+var states = {
+  BUSY: 'BUSY',
+  INUSE: 'INUSE',
+  IDLE: 'NOT_INUSE',
+  RINGING: 'RINGING',
+};
+
+module.exports = states;
