+++ tests/sla_tests.js
@@ -88,7 +88,8 @@ var getMockClient = function() {
           }
         });
         cb(null);
-      }
+      },
+      length: channels.length
     };
     this.deviceStates = {
       update: function(params, cb) {
@@ -183,6 +184,7 @@ var getMockBridge = function(param) {
       });
       cb(null);
     };
+    this.channels = bridgeChannels;
   };
   util.inherits(Bridge, Emitter);
   var mockBridge = new Bridge(param);
